pyproject.toml
[project]
name = "DjangoStarter"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "DealiAxy", email = "dealiaxy@gmail.com"},
]
dependencies = [
    "django[argon2]>=5.0.14",
    "django-ninja>=1.4.1",
    "django-multi-captcha-admin>=2.0.0",
    "django-simple-captcha>=0.6.2",
    "django-cors-headers>=4.7.0",
    "django-ratelimit>=4.1.0",
    "orjson>=3.10.18",
    "pyjwt>=2.10.1",
    "wechatpy>=1.8.18",
    "django-redis>=5.4.0",
    "pycryptodome>=3.22.0",
    "tblib>=3.1.0",
    "jinja2>=3.1.6",
    "faker>=25.9.2",
    "django-split-settings>=1.3.2",
    "django-compressor>=4.5.1",
    "daphne>=4.1.2",
    "django-unfold>=0.57.0",
    "prometheus-client>=0.21.1",
    "requests>=2.32.3",
    "redis>=6.1.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.4",
    "psutil>=5.9.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.1.1",
    "django-debug-toolbar>=4.3.0",
]
